+++ src/web/aws-cloudwatch/KinesisStreams.jsx
@@ -5,7 +5,7 @@ import { Col, Row } from 'react-bootstrap';
 import FormAdvancedOptions from './FormAdvancedOptions';
 import { FormDataContext } from './context/FormData';
 import { ApiContext } from './context/Api';
-import useFetch from './hooks/fetch';
+import useFetch from './hooks/useFetch';
 
 import FormWrap from '../common/FormWrap';
 import ValidatedInput from '../common/ValidatedInput';
