+++ src/web/aws-cloudwatch/StepAuthorize.jsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 
 import { FormDataContext } from './context/FormData';
 import { ApiContext } from './context/Api';
-import useFetch from './hooks/fetch';
+import useFetch from './hooks/useFetch';
 
 import ValidatedInput from '../common/ValidatedInput';
 import FormWrap from '../common/FormWrap';
